Extract ApacheChart utils code
