+++ src/components/ApacheChart/ApacheChart.tsx
@@ -1,64 +1,11 @@
-import deepmerge from 'deepmerge';
-import * as echarts from 'echarts';
 import React, { FunctionComponent, useEffect, useRef } from 'react';
-import { defaultOptions } from '../../utils/echarts';
+import { makeBasicLineChart } from './utils';
 
 type ApacheChartProps = {
   width?: string;
   height?: string;
   options: echarts.EChartOption;
   type?: 'bar' | 'line' | 'pie';
-};
-
-const makeBasicLineChart = (node: HTMLDivElement) => {
-  const chart = echarts.init(node);
-  const option: echarts.EChartOption = {
-    tooltip: {
-      trigger: 'axis',
-    },
-    legend: {
-      data: ['Step Start', 'Step Middle', 'Step End'],
-    },
-    grid: {
-      left: '3%',
-      right: '4%',
-      bottom: '3%',
-      containLabel: true,
-    },
-    toolbox: {
-      feature: {
-        saveAsImage: {},
-      },
-    },
-    xAxis: {
-      type: 'category',
-      data: ['Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat', 'Sun'],
-    },
-    yAxis: {
-      type: 'value',
-    },
-    series: [
-      {
-        name: 'Step Start',
-        type: 'line',
-        step: 'start',
-        data: [120, 132, 101, 134, 90, 230, 210],
-      },
-      {
-        name: 'Step Middle',
-        type: 'line',
-        step: 'middle',
-        data: [220, 282, 201, 234, 290, 430, 410],
-      },
-      {
-        name: 'Step End',
-        type: 'line',
-        step: 'end',
-        data: [450, 432, 401, 454, 590, 530, 510],
-      },
-    ],
-  };
-  chart.setOption(deepmerge(defaultOptions, option));
 };
 
 const ApacheChart: FunctionComponent<ApacheChartProps> = (props) => {
